本登録処理のテストコードを作成
