+++ src/tests/Feature/Auth/RegisterTest.php
@@ -3,7 +3,6 @@
 namespace Tests\Feature\Auth;
 
 use App\Models\Users\TmpUserRegistration;
-use App\Models\Users\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
@@ -22,7 +21,7 @@ class RegisterTest extends TestCase
         //
         $response = $this->withHeaders([
             'XSRF-TOKEN' => csrf_token(),
-        ])->postJson('/user/regist', [
+        ])->postJson('/api/user/regist', [
             'user_name' => 'Sally',
             'email' => 'test3@test.com',
             'birthday' => '2002-11-1',
@@ -48,7 +47,7 @@ class RegisterTest extends TestCase
 
         $response = $this->withHeaders([
             'XSRF-TOKEN' => csrf_token(),
-        ])->postJson('/user/regist', [
+        ])->postJson('/api/user/regist', [
             'user_name' => 'Bakky',
             'email' => 'test3@test.com',
             'birthday' => '1905-1-5',
@@ -64,57 +63,4 @@ class RegisterTest extends TestCase
 
         $response->assertStatus(200)->assertJson(['result' => 'success']);
     }
-
-    /**
-     * すでに仮登録済みのユーザーを新規登録するテスト
-     *
-     * @return void
-     */
-    public function test_create_new_user()
-    {
-
-        $response = $this->withHeaders([
-            'XSRF-TOKEN' => csrf_token(),
-        ])->postJson('/user/regist/complete/testtesttesttest');
-
-        self::assertDatabaseHas(User::class, [
-            'user_name' => 'testNewUser',
-            'email' => 'newtestUser@test.com',
-            'birthday' => '1999-12-1',
-        ]);
-
-        $this->assertDatabaseMissing(TmpUserRegistration::class, [
-            'user_name' => 'testNewUser',
-            'email' => 'newtestUser@test.com',
-            'birthday' => '1999-12-1',
-        ]);
-
-        $response->assertStatus(200)->assertJson(['expire' => true]);
-    }
-
-    /**
-     * すでに仮登録済みのユーザーを新規登録するテスト
-     *
-     * @return void
-     */
-    public function test_create_new_user_fail()
-    {
-
-        $response = $this->withHeaders([
-            'XSRF-TOKEN' => csrf_token(),
-        ])->postJson('/user/regist/complete/failfailfailfail');
-
-        $this->assertDatabaseMissing(TmpUserRegistration::class, [
-            'user_name' => 'testExpieredUser',
-            'email' => 'testExpieredUser@test.com',
-            'birthday' => '1999-12-1',
-        ]);
-
-        $response->assertStatus(404)->assertJson([
-            'message' => '有効期限が切れました。<br>再度登録をしてください。',
-            'expire' => false,
-        ]);
-    }
-
-
 }
